requirements.txt
pandas>=2.0.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
watchdog>=3.0.0
asyncio>=3.4.3
pytz>=2024.1
loguru>=0.7.2
Faker>=20.1.0
tortoise-orm>=0.20.0
aiopg>=1.4.0
asyncpg>=0.29.0